Move EvaluationThreadMonitor to its own class
