+++ src/main/java/org/polypheny/simpleclient/scenario/EvaluationThreadMonitor.java
@@ -1,0 +1,56 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019-2025 The Polypheny Project
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package org.polypheny.simpleclient.scenario;
+
+import java.util.List;
+import lombok.Getter;
+
+public final class EvaluationThreadMonitor {
+
+    private final List<EvaluationThread> threads;
+    @Getter
+    private Exception exception;
+    @Getter
+    private boolean aborted;
+
+
+    public EvaluationThreadMonitor( List<EvaluationThread> threads ) {
+        this.threads = threads;
+        this.aborted = false;
+    }
+
+
+    public void abortAll() {
+        this.aborted = true;
+        threads.forEach( EvaluationThread::abort );
+    }
+
+
+    public void notifyAboutError( Exception e ) {
+        exception = e;
+        abortAll();
+    }
+
+}
